Refactored integration tests
I needed to create a test that has ProcessManager but none of the git
initialization, I realized that the oo hierarchy could be restructured.
