+++ src/tests/IntegrationTests/IntegrationTests.csproj
@@ -76,7 +76,7 @@
     <Compile Include="Events\RepositoryManagerTests.cs" />
     <Compile Include="Events\RepositoryWatcherTests.cs" />
     <Compile Include="BaseGitEnvironmentTest.cs" />
-    <Compile Include="BaseGitRepoTest.cs" />
+    <Compile Include="BasePlatformIntegrationTest.cs" />
     <Compile Include="BaseIntegrationTest.cs" />
     <Compile Include="GitClientTests.cs" />
     <Compile Include="Git\GitSetupTests.cs" />
